clean code and apply script module
