+++ index.html
@@ -57,5 +57,8 @@
     </div>
     <footer>Made by Riuseigi &copy; <span id="year"></span></footer>
   </body>
-  <script src="./scripts/script.js"></script>
+  <script>
+    document.getElementById("year").innerHTML = new Date().getFullYear();
+  </script>
+  <script src="./scripts/navBar.js"></script>
 </html>
