+++ pokedex.html
@@ -64,6 +64,7 @@
       <button id="loadMoreBtn" class="load-more-btn">Load More</button>
     </section>
   </body>
-  <script src="./scripts/pokedex.js"></script>
+  <script type="module" src="/scripts/pokedex.js"></script>
+  <script src="./scripts/navBar.js"></script>
   <script src="https://unpkg.com/aos@2.3.1/dist/aos.js"></script>
 </html>
