+++ scripts/displayErrorMessage.js
@@ -1,0 +1,6 @@
+export const displayErrorImage =() => {
+    pokemonContainer.innerHTML = `<img src="./img/teamRocket.png" alt="" class="errorImage">
+     `
+     loadMoreBtn.style.display = "none";
+   }
+
