+++ scripts/navBar.js
@@ -1,4 +1,4 @@
-document.getElementById("year").innerHTML = new Date().getFullYear();
+//function for the navbar
 document.querySelector('.hamburger').addEventListener('click', function() {
     document.querySelector('.navLinks').classList.toggle('active');
     document.querySelector('.hamburger').classList.toggle('active');
